requirements.txt
nltk
textblob 
newspaper3k                                                                            .env  
